Window modified
front package added.
-controller methods implementeds in window class
-price field added
-error not handled when price or height field is a string.
-stop with the front end for now
