+++ .idea/workspace.xml
@@ -1,8 +1,31 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="BranchesTreeState">
+    <expand>
+      <path>
+        <item name="ROOT" type="e8cecc67:BranchNodeDescriptor" />
+        <item name="LOCAL_ROOT" type="e8cecc67:BranchNodeDescriptor" />
+      </path>
+      <path>
+        <item name="ROOT" type="e8cecc67:BranchNodeDescriptor" />
+        <item name="REMOTE_ROOT" type="e8cecc67:BranchNodeDescriptor" />
+      </path>
+      <path>
+        <item name="ROOT" type="e8cecc67:BranchNodeDescriptor" />
+        <item name="REMOTE_ROOT" type="e8cecc67:BranchNodeDescriptor" />
+        <item name="GROUP_NODE:origin" type="e8cecc67:BranchNodeDescriptor" />
+      </path>
+    </expand>
+    <select />
+  </component>
   <component name="ChangeListManager">
     <list default="true" id="e4a9078e-5ce4-46d7-86d3-a44c2b9c4e72" name="Default Changelist" comment="">
-      <change beforePath="$PROJECT_DIR$/FlowerShop/src/com/crud/domain/Business.java" beforeDir="false" afterPath="$PROJECT_DIR$/FlowerShop/src/com/crud/domain/Business.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/FlowerShop/.idea/.gitignore" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/FlowerShop/.idea/uiDesigner.xml" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/FlowerShop/src/com/crud/domain/Tree.java" beforeDir="false" afterPath="$PROJECT_DIR$/FlowerShop/src/com/crud/domain/Tree.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/FlowerShop/src/com/crud/domain/Window.java" beforeDir="false" afterPath="$PROJECT_DIR$/FlowerShop/src/com/front/domain/Window.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/FlowerShop/src/com/crud/view/MainApp.java" beforeDir="false" afterPath="$PROJECT_DIR$/FlowerShop/src/com/front/view/MainApp.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -19,7 +42,13 @@
   </component>
   <component name="PropertiesComponent">
     <property name="RunOnceActivity.OpenProjectViewOnStart" value="true" />
-    <property name="last_opened_file_path" value="$PROJECT_DIR$" />
+    <property name="last_opened_file_path" value="$USER_HOME$/IdeaProjects/GitProyects/M10_FlowerShop-backend1" />
+  </component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\piero\Documents\GitHub\M10_FlowerShop\FlowerShop\src\com\front\domain" />
+      <recent name="C:\Users\piero\Documents\GitHub\M10_FlowerShop\FlowerShop\src\com\front\view" />
+    </key>
   </component>
   <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
   <component name="TaskManager">
@@ -32,14 +61,26 @@
     </task>
     <servers />
   </component>
+  <component name="Vcs.Log.Tabs.Properties">
+    <option name="TAB_STATES">
+      <map>
+        <entry key="MAIN">
+          <value>
+            <State />
+          </value>
+        </entry>
+      </map>
+    </option>
+    <option name="oldMeFiltersMigrated" value="true" />
+  </component>
   <component name="WindowStateProjectService">
-    <state x="421" y="58" key="FileChooserDialogImpl" timestamp="1601280707752">
+    <state x="421" y="58" key="FileChooserDialogImpl" timestamp="1601374033261">
       <screen x="0" y="0" width="1280" height="680" />
     </state>
-    <state x="421" y="58" key="FileChooserDialogImpl/0.0.1280.680@0.0.1280.680" timestamp="1601280707752" />
-    <state x="339" y="101" key="NewModule_or_Project.wizard" timestamp="1601280723249">
+    <state x="421" y="58" key="FileChooserDialogImpl/0.0.1280.680@0.0.1280.680" timestamp="1601374033261" />
+    <state x="339" y="101" key="NewModule_or_Project.wizard" timestamp="1601372690870">
       <screen x="0" y="0" width="1280" height="680" />
     </state>
-    <state x="339" y="101" key="NewModule_or_Project.wizard/0.0.1280.680@0.0.1280.680" timestamp="1601280723249" />
+    <state x="339" y="101" key="NewModule_or_Project.wizard/0.0.1280.680@0.0.1280.680" timestamp="1601372690870" />
   </component>
 </project>
